+++ yangvalidator/yangvalidator/v2/illegalMethodError.py
@@ -17,10 +17,12 @@ __copyright__ = "Copyright The IETF Trust 2021, All Rights Reserved"
 __license__ = "Apache License, Version 2.0"
 __email__ = "miroslav.kovac@pantheon.tech"
 
-from django.urls import path, include
 
-app_name = 'yangvalidator'
-urlpatterns = [
-    path('yangvalidator/', include('yangvalidator.v1.urls')),
-    path('yangvalidator/v2/', include('yangvalidator.v2.urls'))
-]
+class IllegalMethodError(Exception):
+
+    def __init__(self, used_method: str, *args: object) -> None:
+        super().__init__(*args)
+        self.__used_method = used_method
+
+    def __str__(self) -> str:
+        return 'Only POST request allowed but {} found'.format(self.__used_method)