+++ CHANGELOG.md
@@ -1,6 +1,11 @@
-## Yangvalidator Release Notes
+## Yang Validator Release Notes
 
 * ##### vm.m.p - 2021-MM-DD
+
+  * Make it possible to create request to validate modules outside of webpage [#39](https://github.com/YangCatalog/bottle-yang-extractor-validator/issues/39)
+  * Give options to choose which file to use as a dependency [#9](https://github.com/YangCatalog/bottle-yang-extractor-validator/issues/9)
+  * Create new endpoints to validate modules fo new UI [#48](https://github.com/YangCatalog/bottle-yang-extractor-validator/issues/48)
+  * Upgrade django file structure
 
 * ##### v3.2.1 - 2021-05-04
 
@@ -23,7 +28,7 @@
 
   * Fixed /api/versions endpoint [#45](https://github.com/YangCatalog/bottle-yang-extractor-validator/issues/45)
   * ConfD update [deployment #34](https://github.com/YangCatalog/deployment/issues/34)
-  * Pyang validator update  [deployment #36]( https://github.com/YangCatalog/deployment/issues/36)
+  * Pyang validator update [deployment #36](https://github.com/YangCatalog/deployment/issues/36)
   * YumaPro validator update
   * Fixed validation of a module/s in a .zip file [#40](https://github.com/YangCatalog/bottle-yang-extractor-validator/issues/40)
   * Update of xym tool